backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings
passlib[argon2]
argon2-cffi
python-jose[cryptography]
python-multipart
email-validator



